Add IndexOf method for searching in an array
Implemented a new `IndexOf` method in the `CSharpConcepts` namespace within `Arrays.cs`. This method allows users to search for a number in a predefined integer array and outputs the position if found.

Updated `Program.cs` to comment out the `Arrays.ClearArray()` call and uncomment the `Arrays.IndexOf()` method, enabling the search functionality in the program.